Split Work item specific metadata to own package to import in go-common.
