+++ work/service/emailnotificationsprocessor/claim_account.go
@@ -6,21 +6,13 @@ import (
 	"time"
 
 	"github.com/tidepool-org/platform/pointer"
-	"github.com/tidepool-org/platform/structure"
 	"github.com/tidepool-org/platform/work"
+	. "github.com/tidepool-org/platform/work/service/emailnotificationsprocessor/metadata"
 )
 
 const (
 	claimAccountProcessorType = "org.tidepool.processors.claim.account"
 )
-
-// ClaimAccountReminderData is the metadata added to a [work.Work] item for reminding users to claim their account.
-type ClaimAccountReminderData struct {
-	UserId     string
-	Email      string
-	ClinicId   string
-	ClinicName string
-}
 
 type claimAccountProcessor struct {
 	dependencies Dependencies
@@ -63,20 +55,6 @@ func (p *claimAccountProcessor) Process(ctx context.Context, wrk *work.Work, upd
 		return NewFailedResult(err, wrk)
 	}
 	return *work.NewProcessResultDelete()
-}
-
-func (d *ClaimAccountReminderData) Parse(parser structure.ObjectParser) {
-	d.UserId = *parser.String("userId")
-	d.Email = *parser.String("email")
-	d.ClinicId = *parser.String("clinicId")
-	d.ClinicName = *parser.String("clinicName")
-}
-
-func (d *ClaimAccountReminderData) Validate(validator structure.Validator) {
-	validator.String("userId", &d.UserId).NotEmpty()
-	validator.String("email", &d.Email).NotEmpty()
-	validator.String("clinicId", &d.ClinicId).NotEmpty()
-	validator.String("clinicName", &d.ClinicName).NotEmpty()
 }
 
 func NewClaimAccountWorkCreate(notBefore time.Time, metadata ClaimAccountReminderData) *work.Create {
